requirements.txt
openai
neo4j
numpy
python-dotenv
